🧪: Finish all adapter tests.
