--- src/test/java/com/github/equal/processor/adapter/TimeAdapterTest.java
@@ -1,71 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) [2019] [He Zhang]
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished
- *  to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.github.equal.processor.adapter;
-
-import org.junit.jupiter.api.Test;
-
-import java.time.LocalTime;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class TimeAdapterTest {
-
-    @Test
-    void fromStringHMS() {
-        String timePattern = "HH:mm:ss";
-        String time = "13:02:04";
-        TimeAdapter timeAdapter = new TimeAdapter(timePattern);
-        LocalTime localTime = timeAdapter.fromString(time);
-
-        assertEquals(localTime.getHour(), 13);
-        assertEquals(localTime.getMinute(), 2);
-        assertEquals(localTime.getSecond(), 4);
-    }
-
-    @Test
-    void fromStringHM() {
-        String timePattern = "HH:mm";
-        String time = "13:02";
-        TimeAdapter timeAdapter = new TimeAdapter(timePattern);
-        LocalTime localTime = timeAdapter.fromString(time);
-
-        assertEquals(localTime.getHour(), 13);
-        assertEquals(localTime.getMinute(), 2);
-        assertEquals(localTime.getSecond(), 0);
-    }
-
-    @Test
-    void fromStringH() {
-        String timePattern = "HH";
-        String time = "13";
-        TimeAdapter timeAdapter = new TimeAdapter(timePattern);
-        LocalTime localTime = timeAdapter.fromString(time);
-
-        assertEquals(localTime.getHour(), 13);
-        assertEquals(localTime.getMinute(), 0);
-        assertEquals(localTime.getSecond(), 0);
-    }
-
-}
